help compilation on grid5000 with the paranoid compilation mode...
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@1729 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ src/xbt/config.c
@@ -541,8 +541,8 @@ xbt_cfg_set_parse(xbt_cfg_t cfg, const char *options) {
       option=NULL; /* don't do next iteration */
 
     } else {
-      DEBUG3("Boundary on '%c'. len=%d;option-name=%d",
-	     *option,len,option-name);
+      DEBUG3("Boundary on '%c'. len=%d;option-name=%ld",
+	     *option,len,(long)(option-name));
 
       /* Pass the following blank chars */
       *(option++)='\0';
